Refactor the test file structure
