--- targettests/Remote-Sim/args_synthesis_test_cases/1_single_arg_tests.h
@@ -1,41 +1,0 @@
-/****************************************************************-*- C++ -*-****
- * Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                  *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-#pragma once
-
-#include "util_macros.h"
-
-//=============================================================================//
-//
-// Single argument type tests
-//
-//=============================================================================//
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__), (bool, a, BOOL_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__), (char, a, CHAR_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__), (short, a, SHORT_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__), (int, a, INT_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::size_t, a, SIZE_T_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (int64_t, a, INT64_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (double, a, DOUBLE_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__), (float, a, FLOAT_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<bool>, a, VEC_BOOL_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<char>, a, VEC_CHAR_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<int>, a, VEC_INT_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<std::size_t>, a, VEC_SIZE_T_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<double>, a, VEC_DOUBLE_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<float>, a, VEC_FLOAT_VAL));
-DEFINE_TEST_KERNEL(MACRO_CONCAT(arg_test_, __COUNTER__),
-                   (std::vector<short>, a, VEC_SHORT_VAL));
